requirements.txt
# Install these requirements if you wish to contribute to this project.

# ===========================================================================
# Packages essential to this app. Needed by anyone who wants to use this app.
# ===========================================================================
Django==1.4.5
django-cms==2.3.5
Pillow==2.0.0
django-filer==0.9.3
South

# ==============================================================
# Packages needed for running the tests. Needed by contributors.
# ==============================================================
factory_boy
flake8
nose
django-nose
coverage
django-coverage
ipdb
